LayoutTests/imported/w3c/web-platform-tests/speculation-rules/speculation-tags/same-site-to-cross-site-redirection-prefetch.https-expected.txt


FAIL same-site to cross-site redirection prefetch should not have Sec-Speculation-Tags be sent assert_equals: expected "__no_tags__" but got "tag1"
